Only log ArgumentErrors
This changes the HorizonReader class so that those
errors will not result in indexing stopping.

We have encountered a large number of UTF-8 errors
with recent MARC loads into Horizon. This will
ensure that the bib number is logged so that they can
be corrected, but won't stop indexing completely.

Other Exceptions will cause indexing to stop.